cloud commit

+++ src/main/java/post/Service/CommentService.java
@@ -1,16 +1,21 @@
 package post.Service;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
-import post.DTO.CommentResponse;
+import post.APIResponse.APIResponse;
+import post.DTO.CommentDTO;
+import post.Responses.CommentResponse;
 import post.Entities.Comment;
 import post.Entities.Post;
 import post.Entities.UserFriend;
+import post.Entities.UserProfile;
+import post.Enum.FriendshipStatus;
 import post.Repositories.CommentRepository;
 import post.Repositories.PostRepository;
 import post.Repositories.UserFriendRepository;
+import post.Repositories.UserProfileRepository;
+import post.Security.UserIdContextHolder;
 
 import java.time.LocalDateTime;
 import java.util.ArrayList;
@@ -25,38 +30,53 @@ public class CommentService {
     private PostRepository postRepository;
     @Autowired
     private UserFriendRepository userFriendRepository;
+    @Autowired
+    private UserProfileRepository userProfileRepository;
 
-    public ResponseEntity<?> writeACommentForPost(Comment comment) {
-        int postId = comment.getPost().getId();
-        int userId=comment.getUser().getId();
+    public ResponseEntity<APIResponse> writeACommentForPost(CommentDTO commentDTO) {
+        int postId = commentDTO.getPostId();
+        int userId = UserIdContextHolder.getUserId();
 
-        Optional<Post> postOptional = postRepository.findById(postId);
-        if (postOptional.isEmpty()) {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Post not found.");
+        Optional<Post> optionalSharePost = postRepository.findById(postId);
+        Optional<UserProfile> userProfile = userProfileRepository.findById(userId);
+
+        if (!optionalSharePost.isPresent()) {
+            return APIResponse.error("Post not found.");
         }
-        Post post = postOptional.get();
+
+        if (!userProfile.isPresent()) {
+            return APIResponse.error("User not found.");
+        }
+
+        Post post = optionalSharePost.get();
+        UserProfile user = userProfile.get();
         int postOwnerId = post.getUser().getId();
 
         if (userId == postOwnerId) {
+            Comment comment = new Comment();
+            comment.setUser(user);
+            comment.setPost(post);
+            comment.setComment(commentDTO.getComment());
             comment.setCommentedAt(LocalDateTime.now());
             Comment savedComment = commentRepository.save(comment);
-            ResponseEntity.ok("you are commented on your own post");
-
+            return APIResponse.success("You commented on your own post.", commentDTO);
         } else {
-            // User is not the owner of the post, check if they are friends
             Optional<UserFriend> friend = Optional.ofNullable(userFriendRepository.findBySenderIdAndReceiverId(postOwnerId, userId));
-            if (friend.isEmpty() || !friend.get().getStatus().equals("ACCEPTED")) {
-                return ResponseEntity.status(HttpStatus.FORBIDDEN).body("You can't comment on this post.");
+            if (friend.isEmpty() || friend.get().getStatus() != FriendshipStatus.ACCEPTED) {
+                return APIResponse.error("You can't comment on this post.");
             }
         }
 
+        Comment comment = new Comment();
+        comment.setUser(user);
+        comment.setPost(post);
+        comment.setComment(commentDTO.getComment());
         comment.setCommentedAt(LocalDateTime.now());
         Comment savedComment = commentRepository.save(comment);
-
-        return ResponseEntity.ok("Comment added successfully.");
+        return APIResponse.success("Comment added successfully.", savedComment);
     }
 
-    public ResponseEntity<?> getCommentsByPostId(int postId) {
+    public ResponseEntity<APIResponse> getCommentsByPostId(int postId) {
 
       List<Comment> commentList=  commentRepository.findByPostId(postId);
       List<CommentResponse> commentsList=new ArrayList<>();
@@ -70,23 +90,24 @@ public class CommentService {
               commentsList.add(commentResponse);
           }
       if(commentList.isEmpty()  || commentList==null){
-          return  ResponseEntity.status(HttpStatus.NOT_FOUND).body("this post have no comments");
+          return  APIResponse.error("this post have no comments");
       }
-       return ResponseEntity.ok(commentsList);
+       return APIResponse.success("comments are :",commentsList);
     }
+    public ResponseEntity<APIResponse> deleteCommentById(int commentId, int userId) {
+        Optional<Comment> optionalComment = commentRepository.findById(commentId);
+        if (optionalComment.isEmpty()) {
+            return APIResponse.error("Comment not found.");
+        }
+        Comment comment = optionalComment.get();
+        int commentUserId = comment.getUser().getId();
 
-    public ResponseEntity<?> deleteCommentById(int commentId,int userId) {
-
-       Optional<Comment> comment=commentRepository.findById(commentId);
-        int commentUserId=comment.get().getId();
-        if(userId!=commentUserId){
-            return  ResponseEntity.status(HttpStatus.BAD_REQUEST).body("you are not eligible delete This comment");
-                    }
+        if (userId != commentUserId) {
+            return APIResponse.error("You are not eligible to delete this comment.");
+        }
         commentRepository.deleteById(commentId);
-        return ResponseEntity.ok("comment deleted successfully");
+        return APIResponse.success("Comment deleted successfully", comment);
     }
-
-
 
 
 }
